Add upstream requirement for notify lib
