+++ requirements.txt
@@ -2,6 +2,7 @@ git+https://github.com/madzak/python-json-logger.git@v0.1.5#egg=python-json-logg
 git+https://github.com/alphagov/digitalmarketplace-utils.git@24.0.0#egg=digitalmarketplace-utils==24.0.0
 git+https://github.com/alphagov/digitalmarketplace-apiclient.git@7.5.0#egg=digitalmarketplace-apiclient==7.5.0
 git+https://github.com/alphagov/digitalmarketplace-content-loader.git@1.0.0#egg=digitalmarketplace-content-loader==1.0.0
+git+https://github.com/alphagov/notifications-python-client.git@4.0.0#egg=notifications-python-client==4.0.0
 
 unicodecsv==0.14.1
 python-dateutil==2.4.2
