--- tests/helpers.py
@@ -1,14 +1,0 @@
-class FakeMail(object):
-    """An object that equals strings containing all of the given substrings
-
-    Can be used in mock.call comparisons (eg to verify email templates).
-
-    """
-    def __init__(self, *substrings):
-        self.substrings = substrings
-
-    def __eq__(self, other):
-        return all(substring in other for substring in self.substrings)
-
-    def __repr__(self):
-        return "<FakeMail: {}>".format(self.substrings)
